Added the machine view model.
